Add DesignerActionItemCollectionExtensions in new file
